LMA: Use "timeout" when executing "nc -l..."
When using "nc -l" (i.e., netcat's --listen option), we can't specify
a timeout for the operation, which means that netcat would be blocked
waiting for an incoming connection forever (I had this problem during
one of the tests I was doing).  We can use the "timeout" program to
prevent this scenario.

Signed-off-by: Sergio Durigan Junior <sergio.durigan@canonical.com>